+++ CHANGELOG.md
@@ -19,3 +19,10 @@ dependency updates
 ## 1.2.0
 
 mermaid graph validation added
+
+## 1.3.0
+
+dependency updates
+improved chunking:
+* code gets split up based on tokens
+* text content like md files, indexed based on paragraphs